Results are now formatted properly. The phone will use bold and italic fonts if the word/p.o.s. contains no Myanmar text.
Also fixed a drawing bug with ZawgyiComponent. Now, you should be able to draw text anywhere.

We need to segment syllables properly, so that we can line break accurately. This is somewhat difficult in ZG2008. I'll have to import some code. Alternatively, we can just segment in ZG2009, if the dictionary is encoded properly.